add README, move build scripts into own directory + cleanup impl, remove unused dep & files.
