+++ package.json
@@ -5,12 +5,11 @@
   "scripts": {
     "start": "NODE_ENV='development' webpack-dev-server -c webpack.config.js --silent",
     "build": "yarn _transpile-src && NODE_ENV='production' BABEL_ENV='modern' webpack -c webpack.config.js --silent && yarn _build-inline-template",
-    "_build-inline-template": "NODE_ENV='production' ts-node -T --project tsconfig.node.json ./build.ts",
-    "_transpile-src": "NODE_ENV=production node -r esm -r ts-node/register/transpile-only transpile.ts"
+    "_build-inline-template": "NODE_ENV='production' ts-node -T --project tsconfig.node.json ./scripts/build.ts",
+    "_transpile-src": "NODE_ENV=production node -r esm -r ts-node/register/transpile-only ./scripts/transpile.ts"
   },
   "dependencies": {
-    "goober": "^1.7.1",
-    "preact": "^10.3.2"
+    "preact": "^10.3.3"
   },
   "devDependencies": {
     "@babel/core": "^7.8.6",
@@ -21,7 +20,7 @@
     "@chrstntdd/node": "^0.0.8",
     "@types/babel__core": "^7.1.6",
     "@types/cross-spawn": "^6.0.1",
-    "@types/node": "^13.1.8",
+    "@types/node": "^13.7.7",
     "babel-loader": "^8.0.6",
     "babel-plugin-preval": "^4.0.0",
     "clean-webpack-plugin": "^3.0.0",
@@ -47,7 +46,7 @@
     "terser": "^4.6.4",
     "terser-webpack-plugin": "^2.3.5",
     "ts-node": "^8.6.2",
-    "typescript": "^3.8.2",
+    "typescript": "^3.8.3",
     "webpack": "^4.41.6",
     "webpack-cli": "^3.3.11",
     "webpack-dev-server": "^3.10.3",
