+++ scripts/util.ts
